Use ApplicationCommand instead of a Singleton
This lets Sublime do the heavy listing of making sure that there's
only ever one instance of our command.

In this case instead of having a class that acts as the index and a
command to use the index, the application command IS the index and
navigation at the same time.

In order for this to do what we want, it has to lazy-load the
resources because for some reason the __init__ method of an
Application command can't invoke sublime API commands even though it
happens after the API is ready.