Include first line match
For the purposes of making this easier to work with, add a first line
match to the syntax so that it's automatically detected as a help file
when it's opened.

Currently the code is still manually setting the syntax; the change is
to make it easier to actually author the help files and probably
should not be released unless the test is made more robust or the
files have a different extension or something.