Include JSON TOC file and update help a bit
This includes the notion of a (currently unused) JSON TOC file, which
lists all of the possible topics for help as keys and the files that
they are contained in as the values.

This will be used in the future to set up so that you only specify a
topic and the help command will determine the file needed and update
it as needed.