Include new single command help
This makes the current single application command the entire hub for
help.

The command works in a few different ways:

    - With a topic passed in, the appropriate file for that topic is
      located and loaded (if it is not already visible), and then the
      target is jumped to.

    - With no topic passed in, if the cursor location in the current
      view is a link, the topic is pulled from that link and the
      command acts as if you passed the topic in directly.

    - When there is no topic passed in and one cannot be gathered from
      the current view, or when the toc argument is True, the command
      shows a quick panel of all of the topics instead, and choosing
      one displays the appropriate help.

As the convention in the help files is that the first thing in any
file is a header with a link target that matches the name of the file,
it is possible to display any file directly (and focus the start of
it) by passing its name in as a topic.

The key bindings have been updated accordingly to only use the new
single command, plus a binding on the question mark key will display
the TOC file.

The command file has been constrained to just show the help index
since there is now TOC support.