New hacky fix for view update problem
With help from my compatriot @KingKeith, the workaround I had in place
for the Sublime bug not visually updating selection changes (i.e.
faking a left key press, which would clear the selection) has been
replaced with a better way to work around the issue.

Here the solution is to add and immediately remove an empty region in
the buffer. Removing a region seems to cause Sublime to flag the
buffer for a visual refresh even though it doesn't actually make any
visual changes.