Change up the help format a bit
In order to make things nice on authors and to remove some redundancy,
the help.json file now takes keys that are help files, and the values
are arrays of help topics.

At load time, the file is exploded out so that the keys are the topics
and the values are the help files, as previous. This makes life easier
since most of the time you want to look up a topic and not a file.

This allows for an easier overview of the help topics as well.