Enhanced help TOC system
Here the index.json file has been fleshed out so that instead of being
a simple list of topics, each topic can be a dictionary which also
contains some caption text.

In use, when the TOC is displayed, the quick panel shows the caption
first, followed by the topic on the second line of the entry, allowing
the TOC to show more meaningful data.

Internally, the structure of the data is changed. The topic dictionary
now keys dictionaries which point out the caption for use in the TOC
and the help file for use in actual link navigation.

This code looks pretty gnarly and could probably be cleaned up a fair
bit more, or at least better documented. It's good enough as a proof
of concept, though.