Tweak the separator class
In order to be of more use, the separator class can now consist of
either '-' or '=' (or a combination of the two of them) at any point
in the line instead of just at the beginning. For now the rule is that
at least 4 of them need to be there.