Remove restriction for single selection on nav
This removes the part of the key binding for skipping between tags so
that it will work even when there is a multiple selection. When this
is the case, the first selection is the one used, and when the link is
navigated to, that link becomes the only selection so it's all gravy.