Include mouse help functionality
This implements an event listener that checks to see when the user
has double clicked on text to select it. If the text being clicked on
is a link in a help document, the link is followed instead of
selecting the text.

This allows for a sort of mouse navigation if desired.