Tighten up the help syntax
This tightens up the help syntax somewhat, using variables for things
that are repeated (where possible) and allowing for vertical pipes to
appear outside of links.

In addition, spaces are no longer allowed in link targets or link text
because that makes the current (simplistic) grammar unable to detect
when vertical pipes and single quotes are used outside of links.