Allow links to package files
This includes the ability to link not only to an external URL but also
to a package file, in the form of Packages/PackageName/package.file.

In the latter case, the internal open_file command is used, so if the
user has no override in place for that package file, the version from
the package file will be loaded as a read-only file instead.

In addition, the index.json file was enhanced so that it is possible
to specify the boolean value True for a help file key to indicate that
the link to it should be the same as the help file, as well as a
shortcut for specifying a string as the value to set only a single
target.

This allows for things like links to package files and URL's to be
easily linked "bare" if desired.

THe how.txt file was updated to include how this works.