Allow navigation to stop at targets too
In order to make it a little nicer to navigate through a help file,
this change modifies the handling of Tab and Shift+Tab so that the
cursor will jump not only to links but also link targets.

In a help file that include multiple sections each flagged with their
own target, this allows you to easily skip to targets as well, making
navigation through the file a little nicer.