Enhance the table of contents
This makes the table of contents code a little nicer by making sure
that when the index.json file is loaded we keep the "topic" entry in
the dictionaries that we get and add one in for the items that we
create.

With the entry directly in the topic lookup object like this, even
though it is redundant is greatly simplifies the code that is
responsible for calculating the captions of the text in the quick
panel.

With some extra modifications to how the toc object is structured we
could easily create a nested table of contents.