Swap how the TOC works
This changes how things work for the automatic TOC generation to make
way for a hierarchical TOC structure.

Instead of assuming that the __toc key will just specify a list of
topics and the captions will come from the definitions in the file,
instead the __toc needs to be a list of full structures.

Functionally everything is the same except more work is done in the
code that is generating the toc.