Enhance help settings
This adds extra settings for the Help syntax so that the help looks
nicer while being viewed.

All options which could affect the visual display of the help are thus
defaulted to a "good" value, which may or may not be overriding a user
setting but which may also be overridden by the user if desired.