Spitball some possible Copyright attributions
This includes a couple of notions; a header copyright notice and a
footer notice (one is on each page) as an attribution of where the
original data came from.

The current footer version also includes a link to the official
documentation. I think regardless of how this shakes one, that link
should exist (and maybe be linked from the header in a non-bare
form?).

Additionally, spacing between sections in the general page have been
increased to stop crowding.