Modify all key bindings to make editing easier
This modifies the key bindings that are used by the help system so
that they are only enabled when a setting _hh_edit is False on the
view. This is always the case except for when you manually set the
setting.

This makes it much easier to edit help text with syntax highlighting
turned on.