Add key binding context for help authoring
This replaces the previous code which required setting a specific
setting on a view in order for the key binding to detect if the help
file in question was being edited or not with a context instead.

The context functions by checking to see if the view is currently
read-only or not so that manually opened files are automatically
considered being "in authoring mode" while the help files as  opened
by the commands are not.