Rename index.json to hyperhelp.json
This modifies the help system so that the metadata file is named
hyperhelp.json instead of index.json and it is stored in the root of
the package that is to provide help.

The current plan is that a package that wants to support help has to
specifically tell the help system that it wants to by invoking a
simple API endpoint, and that endpoint would check the presence of
this file to know if this is a supported package or not.