Switch back to the HyperHelp Help in this package
This shifts the help files from what they were merged from in the
api_doc branch (which is now a separate repository) back to the help
for HyperHelp itself.

This still needs much work since the scope of things is changed, and
this will actually be released as something for other packages to use
as well instead of being a simple tech demo which has been fleshed out
more for OverrideAudit.