Turn the package into a dependency
For development purposes, this sets everything up to be a dependency,
as it needs to be going forward.

For now all this changes is the structure of the contents of the
package, with the addition of some loader code cribbed from
OverrideAudit to allow for the contents to be reloaded during
development to make life easier.