Code cleanup: Support multiple packages
Support is currently crude, but now the command takes a parameter that
tells it what package it should use the help index for, and uses the
table of contents and help files from that package.

This removes the notion that the data comes from the package the code
is currently sitting in (which was vaguely hacky) and includes the
notion that the current help view knows not only what help file it is
currently storing, but also what package it came from.