Rename commands for consistency
This simply changes the commands from a `help` prefix to a
`hyper_help` prefix instead, since that is the source of them and that
helps with namespacing (at least in my head).

Additionally the command to move between links has been renamed to
better note how it's going to be used for navigation and not just
between links (e.g. a forward/back mechanism as well).