Expand navigation system
This modifies the command that does navigation so that it can be
easily expanded in the future.

In particular even though it now only allows for navigation between
links and targets, it takes a command that specifies the navigation
type, allowing us to future expand it to include not only navigation
between links and targets both (as it does now) but also each
individually as well as navigating forward and backward through help
history.