Implement help discovery
This is the start of the change by which we discover all packages that
don't already have a help index loaded and load them. For now this
happens on every command invocation, and quickly finds all of the help
files not already in the list of loaded help files.

This will need more work, but it's good enough for now.

This relaxes the restriction that the hyperhelp.json needs to be in
the root of the package, so that it just needs to exist IN the package
(it should exist only once) by using sublime.find_resources() to
quickly do the scan.