Implement "List Available Help"
This is a continuation of the previous commit, in which we now respond
to there being no package by showing a list of all packages that are
currently supplying help.

Choosing a package from the list causes the command to re-execute
with the given package and toc set to true, so that the table of
contents is displayed for that package.