Include a meta description for use in help
This fleshes out the previous commit by allowing the hyperhelp.json
file to include a description instead of just hard-coding it.

Additionally this also swaps the order of the two fields so that the
prime focus is on the description of what the help is for and not the
package that provides it. This may be changed in the future.