Finalize the first pass of multi package support
This includes the final part of the changes needed to support multiple
help packages.

The help command now assumes that if you give it no explicit arguments
at all (no package, no topic AND toc is set to False), that it should
prompt you for one of the available packages and then re-invoke itself
to display the toc.

If that's not the case, when there is no package set it is either
pulled from the current help view if possible, or the user is
prompted.

From here if toc is set, the table of contents is displayed, otherwise
the given help topic is displayed. In this case if the help topic is
not provided, one is pulled from the current help buffer if it has the
focus and the cursor is on it, otherwise a default of index.txt is
used instead.

With these changes in place, everything is more or less set up as we
want.