Alter the visibility of the TOC command
This disables the help command when it is explicitly told to display
the TOC but there is not an explicit package provided and there is no
currently open help view in the current window from which one could be
gathered.

In such a situation, the command would respond by prompting you for  a
package first, which is not what is being asked (set toc to false if
you want that).

This has the effect of hiding the command from the command palette and
disabling it within menus as two visual indications that the command
is not currently available.