Fix navigation in help files
This fixes a regression introduced in a previous commit in which
pressing enter or double clicking on a link would open the list of all
help packages instead of following the link.

The reason is that these actions both assume that invoking the command
with no arguments will have it follow the link under the cursor, but
that is no longer what happens.

To resolve this, the navigate command now takes a new nav parameter of
"follow" to indicate that it should follow the link under the cursor.

This is something that I wanted to do at a later date, so this just
speeds up the timetable.