Swap order of package name and description
This swaps the name and description fields around again in the list of
packages with help.

Although the previous way is more descriptive, this way mimics how
packages are displayed everywhere else (PackageControl for example).
This presumes that the user is most aware of the packages that they're
using, and secondarily can find what they want by description or
filtering if needed.