Support document root specification in index json
This adds a new optional meta entry in the hyperhelp.json named
"__root" which specifies the root of the documentation provided for
this package.

When it's not given, it defaults to the same location as the
hyperhelp.json file (so e.g. "index.txt" and "hyperhelp.json" would be
in the same package folder).

If it is provided, it's assumed to be a path relative to the root of
the package that contains the hyperhelp.json file. In this case,
"doc", "/doc" and "/doc/" are all equivalent and represent a folder in
the top level of the current package named "doc".