Include developer option to reload help files
This includes one extra parameter to the hyper_help command for
reloading help indexes or help topics that have already been loaded.

The command now has a reload argument which is set to False by
default. Once the potential load of help indexes has been completed,
if this value is set to True then either a reload of help index
information or the currently loaded help topic is done and the command
exits.

If the topic is set to "reload", then the help file currently open in
the help window (if any) is reloaded from disk and placed back at the
top of the file. This allows changes made during authoring to be
realized easier. This can be extended in the future to try and restore
the position of the buffer as well, but this is good enough for now.

If topic is None or otherwise not "reload", then help index
information is loaded instead. The extent of the reload is based on
the value of the package parameter.

If a package is given and it's a package that has already been loaded
previously, the loaded help index for that package is discarded and
it is reloaded. This would bring things like the document root,
description and toc up to date, along with altering the list of known
topics.

The package parameter can also be None, in which case all current help
indexes are discarded and a full fresh reload is done as if the
package was just loaded.

This means that this will effectively reload existing packages, find
any new packages not previously available, and discard information for
packages that no longer exist.

Probably the code should be modified to detect when the list of
ignored packages changes as a consequence of PackageControl
installing, updating or removing a package and invoke the command as
needed to  keep the help system up to date.