Clean up loader code and install instructions
This commit changes the name of loader.py to be loader.code instead,
which is functionally identical in normal operation but has the
advantage that Sublime will not attempt to load the file while it's
scanning packages, which helps to ensure that the package doesn't get
double loaded.

This is due to some comments made by Will Bond (the Package Control
Author) which seems to indicate that this is the preferred method for
this very reason.

Additionally, since there is a bug in the recent versions of Package
Control with regards to installing a local dependency, the READAME now
has more detailed instructions on how exactly to go about performing
the manual install of the package.

In the future, if/when Package Control fixes the problem, the extra
steps can be redacted from the README.