Include a package version number
This includes in the core package the concept of a version number that
underlying code could use to determine what version of the core it is
interfacing with.

This is only really needed for packages that get at the underlying
core code in hyperhelp, which at the moment would only be the
companion HyperHelpAuthor package.

To this end (as an experiment) the code for loading in the help index
files is exposed to anyone that wants to import it as a test for how
this will eventually work. In particular, the idea is to put all of
the symbols meant to be accessible to outside code into the
hyperhelp.all module namespace (whicn is unfortunate but there seems
to be no satisfactory way around it).