API Cleanup
This cleans up some of the API, now that we're moving more fully
towards something that is going to be used by others and not just be
posted as a demo and more fully polished in OverrideAudit.

This is the first stage of the cleanup, where existing methods in the
core operations file are renamed and cleaned up somewhat to be more
useful in the general case, and exposed to users of the module.

More of this will come in upcoming commits where code is
systematically extracted from the core help command file and placed
here (where possible and sensible) to make that code cleaner to look
at.

In particular the ApplicationCommand for help is vaguely unwieldy
at the moment.