Push more help into operations from help
This shifts more of the responsibility of how to find a help view and
load a help file into a reusable state in the operations file.

A bit more work needs to be done to get all of the code out of the
single command class (maybe a couple of more commits) but this is a
good place to start laying blame for things going wrong later if
anything regresses.