Push loading and reloading topics to the API
This shifts the methods in the HyperHelp command class responsible for
showing help topics and reloading the currently visible help file out
to public API functions instead.

Along the way a little bit of cleanup has happened in the help command
to make the code a little bit less hairy.