Update the default syntax
This updates the syntax file to use a different top level scope and
include the notion that there is a special header line that can be
used to recognize that a file is a help file being authored.

The header line supports key/value pairs that are used to provide
metadata to the help system (this is not implemented yet).

As a result of this, and the new plans for a specific help authoring
system, the syntax is hidden. The new syntax will recognize a file
that is being authored, and the code applies the correct syntax to a
help file when opening it.