Shift the help index into the documentation folder
I think in the past I had decided that this would be outside of the
documentation folder as an example of how that would work (I forget
now), but I like the cleanliness of all of the help related files
being contained directly in the doc folder.

This commit shifts the location of the file, while also removing the
explicit document root that it contained, since that will now
automatically be set.