Extend HelpData to include the list of files
This allows external code to easily get at the list of help files
contained in any particular packages' help index file.

The file property excludes all meta files that represent URL's to open
in a browser or local package resource files to directly open in the
editor, and the names are relative to the document root in the help
data itself.