Move sentinel key for scanning out of scan code
In the original prototype code, the help command would try to lazy
load the information for all help indexes only when it was needed
(e.g. when you try to list all available help) and otherwise demand
load the help indexes for packages as they were referenced.

In order to facilitate knowing if a complete scan was needed or not,
the code that does the full scan would include a sentinel key that the
code could check to see if it needed to do a full scan or not.

This was changed a while ago, and while doing this is harmless, it is
annoying for external code that just wants to walk the packages
without having to exclude the key themselves. As such, this commit
makes the help command insert the key itself when it needs to.