Refactor how help indexes work
This makes several enhancements to the HelpData tuple that contains
help data and in the structure of the hyperhelp.json index file to
make authoring tools and other outside code easier to develop.

This includes the notion that the HelpData can tell you not only what
help files currently exist, but also what URLS are being linked to and
what local package files are being linked.

This is done without having to use regular expressions, by making the
index specifically mention that this is how things work.