Minor code cleanup
This cleans up the construction of the HelpData tuple that contains
the index so that the index property becomes index_file to better
indicate what it is, and to ensure that the urls property is a list.

Additionally small whitespace formatting tweaks in the help.py file
that were upsetting my delicate sensibilities.