Detect duplicate help topics
The code responsible for creating the master list of topics for a
particular help file has been slightly modified to report when a help
topic is listed twice by generating a log message in the console.

When this occurs, the first version of the topic seen is the one that
is used, which is probably entirely non-deterministic assuming that
the key traversal of a python dictionary is not in any particular
order, which I presume it is.