Detect missing help topics in the help TOC
The code that loads the help index now verifies that all of the topics
that are referenced in the provided table of contents actually exist
in the file, and will generate console warnings for any that are
missing to let the help author know that the TOC may not work as
expected.

This check only happens when the help index specifically contains a
table of contents; when it does not, a table of contents is
synthesized by just including all possible help topics, so there is no
need to verify in that case.