Include a setting for in-help navigation type
Previously the Tab/Shift+Tab key combinations would skip the cursor
location to the next or previous link or link target based on the
current cursor position.

This changes the navigation command so that it is now possible to do
this only between links, only between link targets, between both (as
previous) or control which of those is used based on a setting, which
allows for the operation to be both configurable by the user as well
as easily toggled by commands, plugins etc.

The new setting is "hyperhelp_navigate_type", which if not specified
defaults to the old behaviour.