Change how date setting is loaded
Previously this was loaded directly from the user preferences, but
this now comes from the view settings of the help file view instead,
which inherit from the global preferences anyway.

This is with an eye towards a future OA-like extension where a single
entry point would take a setting and return a known default, which
requires that all settings come from the same place.