Refactor loaded help indexes to be global
Instead of the main help command being responsible for tracking the
currently loaded set of help indexes to allow for help navigation,
there is now a top level module function that is responsible for
loading it and handing it out.

This allows code outside of the help command (such as an on_hover
handler) to access the help information.