Use new global help indexes to support popups
Now that the list of help information is not local to the help
command, the event listener can use on_hover to show a simple popup
that shows the caption of a help topic and where it will link to.

The link location is listed as opening an external URL, opening a
package file or linking to another help file.

At the moment this is not as interesting as it could be because most
of the help topics have no caption assigned to them, which causes the
help loader to default it to the topic. The only exceptions are the
external URL links and the links to package files.

The rest of the topics will get appropriate captions when the default
help files are rewritten.