Refactor global help index into public API
This shifts the _get_help_info() global function (recently introduced)
which tracks the set of global help indexes that have been loaded from
being available only to the official help command to being part of the
public API instead.

This is because this information will soon be needed for code in the
help operations as well during the history refactor. Additionally it
helps with the help authoring tools, which can track and affect the
global help state more easily while authoring help.