Introduce HistoryData and view settings for it
This introduces a new named tuple, HistoryData.

This tuple contains a record of the current help view state; the
package and help topic being displayed, what the viewport layout is,
and the state of the first selection.

In use entries of this type are stored in settings in the help view so
that as help is navigated we can track what help topics were visited,
as well as being able to navigate between them and restore what the
viewport layout looked like at the time.

To this end, when a brand new help view is first created, the
display_help function ensures that there is a stub history entry for
the just-loaded help file.