Introduce history tracking for help views
This introduces the handling for the history of help topics navigated
through in the current help view by updating and appending to the list
of history data that is stored in the view.

In order to track history, the following steps are taken:

First, once show_topic() has determined that it might actually be
following a link (i.e. after error checking and determining that the
topic is not a URL or package file), it will update the current
history entry in the view to ensure that the current viewport and
caret position is captured.

Next, once the link is actually navigated to, a new history entry is
appended to the stack which stores the information for the new help
topic.

This action in show_topic() is only taken when a boolean parameter is
set to True to tell it that it should, so that the upcoming code for
navigating through the history can still show topics without
disturbing the history at all.

At this point history is tracked in the view but cannot be navigated.