Include new navigate_history API method
This API method navigates the help view through its history of topics
by pulling either the next or previous history entry out of the help
view, displaying that topic and then updating the record of which
history entry is currently being displayed.

The method checks to ensure that the current history position can be
moved in the direction that is provided before changing anything, and
will update the current position and viewport in the current history
entry before navigating away, so that you can come back to where you
left.

The navigate command that is globally exposed to Sublime now takes an
extra navigation type, "follow_history", which uses this API method
directly to do the navigation, and the the Alt+Left and Alt+Right
keys are bound to perform the navigation.