Truncate old history on new navigation
When the internal _update_history function is told to append a new
entry to the history, it now first checks to see if the history entry
currently set is the last one in the history or not.

If it is, nothing happens. Otherwise, all existing history past this
point is truncated out of the list prior to the new entry being
appended.

The idea is that if you have navigated somewhere within the history of
help and navigate to a new link (even if it's the same place you  went
before), all history after this is now gone.