Enhance visibility of hyper_help_navigate command
This command for navigating around in help has been enhanced so that
it will disable itself when it cannot be currently executed (mostly;
see below).

In particular, the command will always be disabled if there is no help
view currently open in the window or when the navigation item it has
been given is not one of the valid ones.

Aside of this, the command will also disable itself if the navigation
item tells it to follow history, but it's already at the extreme end
of history and can't navigate in that direction.

In addition, a description method has been implemented which will
generate a custom caption for the follow_history nav command so that
it contains the name of the package and file that it will navigate to.

This can be further enhanced, but this is good enough for now. In
particular, the command could disable itself when it's told to follow
a link but the caret is not currently positioned on one.