Redact "option" link targets
This is (was) a specific type of target which was treated identically to
regular link targets, but which used single quotes instead of asterisks
as delimiters and was colored differently.

In the initial prototype code this was something taken from the vim help
style, which allowed for a slightly different style of link for
targeting settings and options (hence the name).

For now the idea of this as a construct is being removed in favor of
just having a standard link target instead and we will see if any
situation arises which would make this be useful.

In particular I worry that this particular syntax makes it impossible
for help text to contain words in single quotes. There are already
enough potential pitfalls with help text with the existing syntax.