Change context name to have package prefix
hyperhelp includes a context for use in key bindings which tracks
whether or not the current file is a help file that is being authored
or not (as based on it's read-only view attribute).

This change modifies that context to have a hyperhelp prefix to avoid
collisions with other packages. The previous name was a holdover from
when the package was still a tech demo.