Fix the description method for the navigate cmd
The description method for the HyperHelpNavigate command is meant to
always return the sublime default caption for the command except for
the instance where the `nav` option is "follow_history", where the
caption has a default that tries to include the topic that would be
navigated to (where possible).

This had two defects in it. Firstly, it would return None (to get the
default) in cases where there was no help view (and thus no history)
OR when the nav item was not "follow_history", which is incorrect
because even with no help window, if the nav is "follow_history" the
custom caption should still be returned.

Additionally, there is either a bug in Sublime or a bug in the
documented API for the `description` method that says that it should
return None to get the default, but although this works it also causes
Sublime to generate an exception message (but no stack trace) to the
console.

Testing shows that returning an empty string (which is what the actual
default implementation in the sublime_plugin.py file does) causes the
correct behaviour without the warning message in the console.