Include a key binding to open the help index
At least for testing and demonstration purposes, it is handy to have a
key binding set up to open the help index without having to open the
panel first.

As with all of the other current key bindings, this is subject to
change, but using the letter `i` for this doesn't seem overly
egregious at the moment since the help view is read-only anyway.