Switch topics to NBSP from Tab
Currently in order to get some form of visible white space into places
where it would otherwise make the text ambiguous (e.g. inq links and
anchors), we used a tab.

This works but is cumbersome and forces the `tab_size` setting to be
set to a single character, which makes indenting a chore.

In lieu of that we now use the non-breaking space character, which is
U+00A0 for this purpose. This offers the same whitespace visibility as
a tab does, but without requiring a tab size.

When the index is loaded, spaces are turned into non-breaking spaces
to store in the topic list, and lookups also do the same for more
natural topic searches.