Modify the default tab settings in help files
As tab characters are no longer special in any way, the default
options for the syntax have been modified so that they don't force tab
characters to be a single character and to ensure that no physical
tabs are inserted for indenting.

This eases the restrictions on tabs existing, but they should still
not be used since there is and will never be any consensus on the
appropriate width, and the help should appear to the viewer the same
as it does for the author.