Catch package renames during load
Currently the package that a help index is for is inferred from the
package that it's contained in and as a result each package may
contain only a single hyperhelp.json help index.

The code is changing to allow for multiple help indexes in a package
to allow for the case of multiple help systems being deployed at once.

This requires that the package be specifically set in the index file
instead of being inferred from the resource file name, which means
that it is possible from one load to the next for the package in a
help index to change.

This change catches the situation where the package associated with an
index changes when it's reloaded to ensure that we correctly discard
the older package index from the list.