Include new required package field in indexes
This includes a new mandatory field in help indexes named `package`,
which specifies the package that this help index is meant to be for.
This information was previously inferred from the package that the
index resource was stored in.

This change allows for multiple help packages to be contained within a
single Sublime package, such as in SnAPI where the goal is to contain
not only the API information but some Sublime help as well.

This requires a change to the core logic of scanning for packages,
which can no longer assume one help index per package and must now
load the index in order to determine what package it's for.

As implemented here it is possible for multiple help indexes to
declare that they are for the same help package. When this occurs,
some core rules are currently followed:

  1. If only one of the help indexes is canonical, that index is
     always chosen.

  2. If two or more help indexes are canonical, both of them and all
     other indexes that declare that they are for the same package are
     ignored.

  3. If none of the help indexes are canonical, the last one loaded
     "wins" and becomes the index used.

Here a help index being canonical means that it both declares that it
is for a particular package and is also contained within that same
package.

A future mechanism is planned for allowing the user to specify on a
package by package basis which index is the one to be considered the
canonical index, to allow for overriding help or breaking ties in
favor of a particular package.