Include some experimental syntax changes
This change allows code blocks to be highlighted in a nicer manner.

In particular, inline blocks apply a punctuation scope to the open and
close characters and a markdown.raw scope to the interior so that the
punctuation can be targeted with special colors but is more
invisible.

In addition, code blocks now offer an optional language specification
and will (for json, python and xml) use those languages.