Fix menu and command palette entries
The most recent set of changes included a change to the name of the
package that contains the core HyperHelp help, from `hyperhelp` to the
more package-appropriate `HyperHelp`.

An unintended side-effect of this is that the menu and command palette
entries for displaying help were not modified, making them effectively
do nothing.